requirements.txt
Flask>=3.0.0
structlog>=24.0.0
gunicorn>=23.0.0
python-pptx>=1.0.2
pdfplumber>=0.11.0
python-dotenv>=1.0.1
supabase>=2.0.0
